refactor(examples): namespace the Cloudflare orchestration host
